+++ projects/server/controller/developer-controller.js
@@ -1,4 +1,4 @@
-const db = require('../config/db.config');
+const db = require('../config/db');
 const Document = db.Document;
 
 function getAttributeName(attribute) {
