+++ projects/server/parser/audit-parser.js
@@ -1,8 +1,8 @@
 const fs = require('fs-promise');
 const request = require('request');
-const config = require('../config/app.config');
+const config = require('../config/app');
 const parserConfig = config.parser;
-const db = require('../config/db.config');
+const db = require('../config/db');
 const Document = db.Document;
 const path = require('path');
 const logger = require('../core/logger');
