+++ projects/server/routers/audit-router.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 
-const controller = require('../controller/auditController');
+const controller = require('../controller/audit-controller');
 router.post('/audit', controller.postAudit);
 router.get('/subsidiaries', controller.getSubsidiaries);
 router.get('/auditStatuses', controller.getAuditStatuses);
