+++ projects/server/routers/document-router.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 
-const controller = require('../controller/documentController');
+const controller = require('../controller/document-controller');
 router.get('/documents', controller.getDocuments);
 router.get('/documentsByType', controller.getDocumentsByType);
 router
