+++ projects/server/routers/event-router.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 
-const controller = require('../controller/eventController');
+const controller = require('../controller/event-controller');
 router.get('/eventTypes', controller.getEventTypes);
 router.get('/logs', controller.getLogs);
 router.post('/error', controller.postError);
